Allow admins to create and use the opening period as templates
Some of the resources might will have same opening hours. It is
more convinient if the admins can create a template for opening
hours and share that template between multiple resources.

If the template used in a resource is modified, it no longer is
a template and should be treated as normal opening hour period
belonging to a resource.

Refs TAM-122