Change reservable_after to only include today
Previously, when a value is set for minimum_days_before_reservation
the resource would be available to reserve after the value of
minimum_days_before_reservation + 1 days. i.e. If the value is
1 the resource is available for reservation after 2 days from now.

However, client want it so that it doesn't add one extra day. i.e.
If today is Monday and the minimum_days_before_reservation is set
to 1, that means the resource should be reservable tomorrow but
not today.

Refs TAM-145