Move the orderline relevant field to OrderLine model from Product
The price, tax_percentage, price_type, price_period makes more
sense to be in OrderLine model instead of Product because when
we want to look at the historic data, these values needs to be
saved in database as the price/tax etc. will be dynamic values.

The same product will have different prices depending on who is
reserving the resource or for what purpose. For this reason a
product can't have a fixed price related fields. Therefore, store
these fields instead in OrderLine model.

Refs TAM-...