Resources: set serializer max_price and min_price to None if free to use
ResourceSerializer fields max_price and min_price return None if
free_to_use is True, even if the resource has pricing information.

Originally the Resource validation would prevent setting max_price and
min_price if free_to_use is True. As these fields are now calculated
with the queryset with_pricing() method it is possible to have values
for these fields even if free to use, so they are checked in the
serializer instead. This is to ensure backward compatibility with
Varaamo, which expects min and max price to always be None if
free_to_use is True.

Another approach might be to automatically set min_price and max_price to None in the
with_pricing() queryset method, but there may be cases when we want to
query for max and min values for any resource whether free to use or
not.

Refs TTVA-35