Payments: skip order validation and processing if no order required
Currently order validation and potential creation will still happen if
payment is required e.g. if resource is free to use or the user has
staff/admin authorization on the resource.

One example where this happens is when creating a new reservation as a
staff user, as Varaamo will not pass order line data to the API, so the
POST call will not validate.

Unit tests have also been updated to ensure tests are run with a free or
paid resource as appropriate.

Refs TTVA-75