Resources: add unit tests and optimize fetching of price list info
Added a cached property Resource.price_list which should return the
pricing information if the resource is not free to use. This is used in
the resource API to fetch user group and event type info more
efficiently.

Unit tests have also been added for resource API and model changes.

Refs TTVA-59