Payments: ensure order preserved when committing update
Currently if "order" is None (i.e. no order passed) when updating a
reservation with PUT, the saved instance will have order=None.

This change ensures that the current Order instance for the reservation
is saved on commit if not present in the request data.

Unit test also updated to check that the email is sent to the user when
a paid reservation is approved.

Refs TTVA-20