Order.objects.update_expired: Confirm paid orders
When trying to expire an order that is waiting for payment but has
actually already been paid in CeePos, confirm the order instead. This
should not happen but there have been cases where Respa thinks the order
has not been paid yet even though it has been. This can happen if e.g. CeePos
does not send the notification of a successful payment and therefore the
state is not updated in Respa. The root cause for such cases should still
be investigated but this at least fixes the issue of paid orders being cancelled.

Refs TTVA-112