Respa admin: remove "include_other_event_type_option" from price list form and modify check to template
As this field should not be editable and is not shown in the form itself
(even hidden) the check is done on the form PriceList instance instead
and the field removed from the form.

In addition, the logic appeared to be incorrect: the "none of the above"
text should be shown if "include_other_event_type_option" is True, not
False.

As this field is default True, it should always be shown when creating
a new non-template based PriceList.

Refs TTVA-115