Django admin: add "include_other_event_type_option" in Django admin form
This field was removed from Respa admin as it was not actively being
used. This had introduced a bug where the field being present in the
form, but not present in POST, automatically set the value to False
every time the form was saved.

This change adds a subclass of the PriceListForm for the Django admin,
explicitly adding in the missing "include_other_event_type_option".

Refs TTVA-99