Respa admin: enable Copy functionality for period templates
When one or more period templates are selected in the dropdown, shows a
"Copy" button. When clicked, the template is copied into a new form in
the "Periods" formset.

Period templates are rendered in the frontend into <template> tags that
can be safely amended and copied.

In addition the select "Lisää yleinen kausi" has been removed, as it
does not appear to have any functionality attached.

Refs TTVA-33