Respa admin: fix INITIAL_FORMS counter bug in period days formset
The bug is caused by Javascript which sets the INITIAL_FORMS counter for
the days to the number of elements, even for a period that hasn't been
added to the database yet. This causes a validation error as the initial
count should be zero in such cases.

In addition a small UI change has been added to indicate which period in
the formset has an error, as it is currently not possible to determine
from the error list shown at the top of the page. Erroneous periods are
shown in the standard error colour in the CSS, and the expanded form
lists errors specific to that period.

Refs TTVA-140