Reservations: send admin email only when reservation confirmed/requested
Instead of sending the "to officials" email when the reservation is
created, this email is sent only on confirmation (i.e. after payment if
required) or when the reservation requires manual confirmation.

We might want to split these two cases into separate email notifications in future to
make it clear what action is required by the admins.

Refs: TTVA-109