Reservations: add invoicing fields
Adds new INVOICE_REQUESTED status, as well as invoice_requested flag to
indicate customer requested invoice on resources where this is
permitted. The invoice_requested_at timestamp should be set when the
reservation is approved for payment (e.g. for reservations requiring
manual confirmation).

In addition a number of company-specific fields added for email, address
etc. for invoicing requirements.

Refs: TTVA-141