CeePos: Use resource name in order line description
Use the resource's name instead of the product's name as the
order line description in the CeePos integration. The order line
description is visible to the orderer when they make the payment
in CeePos.

When the resource name is changed, the related products' names
are not updated automatically. This lead to the order line description
being incorrect in Ceepos in such cases.

Refs TTVA-150