Save invoicing specific fields to reservation
When payment by invoice is requested, we require certain fields to
be present in the reservation data. However, these fields are not
saved to the Reservation object unless the fields are also included
in the metadata set. That is not the case when the customer wants
to show those fields only when requesting payment by invoice.

Make sure the invocing specific fields are required and saved even
if they aren't included in the metadata set, when invoice is requested.

I also had to add them to the data respresentation in the Reservation
serializer so that those fields are included in the reservation data
returned to the client. This ensures the fields are included in PUT
requests, e.g. when approving a reservation.

Refs TTVA-179