Resources: change Resource.objects.free_of_charge() logic
Previously the implementation would consider "free" to be:

1) Has "free_to_use" value True
2) Does not have a price list (internal payments)
3) Does not have min/max default prices (external payments)

A resource that did not match 2) and 3) but still had free_to_use=False
would still be considered "free".

The logic is now that any resource with free_to_use=True is free, and
all others are considered paid resources, even though there is no
attached pricing information.

In addition this commit removes associated unit tests as the use cases
are much simpler.

Refs: TTVA-189