Archive related Product when PriceList is deleted
Products and PricedProducts are created when a PriceList is created
or edited. When a PriceList is deleted, the related PricedProduct is
deleted. However, the Product was not deleted/archived. This caused
issues with there being multiple Products for the same resource. Fix
this by archiving the existing Product when a PriceList is deleted.

Refs STAM-24