Rename people_capacity to people_capacity_lower
People capacity will be split into two fields: `people_capacity_lower`
and `people_capacity_upper` for adding a possibility to define a range
of the space capacity.

Rename the `people_capacity` field to `people_capacity_lower` in the
`Resource` model and keep the data of the original field in this new
field.

Keep the original name `people_capacity` in the API until UI changes are
done.

Refs TTVA-217