Change Ceepos cost center code to be based on the tax percentage
In Respa the tax precentage for same product may vary. Ceepos has a
different product code for each tax percentage, so we need to get the
correct product code based on the tax percentage when sending the order
to Ceepos.

Change Unit's `cost_center_code` field from CharField to JSONField.
This was agreed to be the most straightforward way to implement this
feature now. The field then contains a mapping of tax percentages to
the correct cost center codes.

Refs TTVA-240